feat: implement MachineASousLogic for handling game logic and Firebase data fetching
